Move ICN image decoding to image_tool.cpp
