+++ icn2img.vcxproj
@@ -39,7 +39,7 @@
     <Keyword>Win32Proj</Keyword>
     <WindowsTargetPlatformVersion>10.0.18362.0</WindowsTargetPlatformVersion>
     <ConfigurationType>Application</ConfigurationType>
-    <PlatformToolset>v142</PlatformToolset>
+    <PlatformToolset>v140</PlatformToolset>
     <CharacterSet>Unicode</CharacterSet>
     <Disable-Microsoft-googletest-v140-windesktop-msvcstl-static-rt-dyn>true</Disable-Microsoft-googletest-v140-windesktop-msvcstl-static-rt-dyn>
   </PropertyGroup>
@@ -47,13 +47,13 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug-SDL1|Win32'" Label="Configuration">
     <ConfigurationType>Application</ConfigurationType>
     <UseDebugLibraries>true</UseDebugLibraries>
-    <PlatformToolset>v142</PlatformToolset>
+    <PlatformToolset>v140</PlatformToolset>
     <CharacterSet>MultiByte</CharacterSet>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug-SDL2|Win32'" Label="Configuration">
     <ConfigurationType>Application</ConfigurationType>
     <UseDebugLibraries>true</UseDebugLibraries>
-    <PlatformToolset>v142</PlatformToolset>
+    <PlatformToolset>v140</PlatformToolset>
     <CharacterSet>MultiByte</CharacterSet>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release-SDL1|Win32'" Label="Configuration">
@@ -201,7 +201,6 @@
     <ClCompile Include="src\engine\translations.cpp" />
     <ClCompile Include="src\engine\xmi2mid.cpp" />
     <ClCompile Include="src\engine\zzlib.cpp" />
-    <ClCompile Include="src\fheroes2\agg\decode.cpp" />
     <ClCompile Include="src\tools\icn2img.cpp" />
     <ClCompile Include="src\thirdparty\libsmacker\smacker.c" />
     <ClCompile Include="src\thirdparty\libsmacker\smk_bitstream.c" />
