+++ src/engine/image_tool.h
@@ -26,4 +26,6 @@ namespace fheroes2
 {
     // Save an image into file. 'background' represents palette index from the original palette
     bool Save( const Image & image, const std::string & path, uint8_t background = 23 );
+
+    Sprite decodeICNSprite( const uint8_t * data, uint32_t sizeData, const int32_t width, const int32_t height, const int16_t offsetX, const int16_t offsetY );
 }
